port_visualization_api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
redis
